Refactor into a Sinatra modular app
