--- app.rb
@@ -1,122 +1,0 @@
-require 'dotenv'
-Dotenv.load
-
-require 'pry'
-require 'twilio-ruby'
-require_relative 'models/client'
-require_relative 'models/message'
-
-require 'sinatra'
-require 'sinatra/activerecord'
-require 'sinatra/reloader' if development?
-
-helpers do
-  def protected!
-    return if authorized?
-    headers['WWW-Authenticate'] = 'Basic realm="Restricted Area"'
-    halt 401, "Not authorized\n"
-  end
-
-  def authorized?
-    @auth ||=  Rack::Auth::Basic::Request.new(request.env)
-    @auth.provided? && @auth.basic? && @auth.credentials &&
-    @auth.credentials == [ENV['AUTH_USERNAME'], ENV['AUTH_PASSWORD']]
-  end
-end
-
-get '/' do
-  protected!
-  @clients = Client.all.order(last_name: :asc)
-  erb :index
-end
-
-post '/send' do
-  account_sid = ENV['TWILIO_SID'] # Your account SID from www.twilio.com/console
-  auth_token = ENV['TWILIO_AUTH_TOKEN'] # Your auth Token from www.twilio.com/console
-
-  # scrub out non-digit characters
-  phone_number = "#{params['phone_number'].gsub(/\D/,'')}"
-
-  if phone_number.length != 10
-    redirect to ('/error')
-  end
-
-  @client = Client.find_or_create_by(phone_number: phone_number)
-  @client.first_name = params['first_name'] if @client.first_name.blank?
-  @client.last_name = params['last_name'] if @client.last_name.blank?
-
-  if @client.save
-    @message = Message.new(text: params['message'], client_id: @client.id, outbound: true)
-  end
-
-  if @message.save
-    begin
-      @twilio = Twilio::REST::Client.new(account_sid, auth_token)
-      message = @twilio.messages.create(
-        body: params['message'],
-        to: "+1#{@client.phone_number}",
-        from: ENV['TWILIO_NUMBER'] # Your Twilio number
-      )
-      puts "Message sent from #{message.sid}"
-    rescue Twilio::REST::TwilioError => e
-      puts e.message
-      redirect to ('/error')
-    end
-    redirect to ("/clients/#{@client.id}")
-  else
-    redirect to ('/error')
-  end
-end
-
-post '/receive' do
-  phone_number = params['From'].gsub('+1','')
-  @client = Client.find_by(phone_number: phone_number)
-  if @client.present?
-    @message = Message.create(text: params['Body'], client_id: @client.id, inbound: true)
-  end
-end
-
-get ('/clients/:client_id') do
-  protected!
-  @client = Client.find(params[:client_id])
-  @client.messages.where(read_at: nil).update_all(read_at: Time.now)
-  @messages = @client.messages.order(created_at: :asc).group_by { |m| m.date }
-
-  erb :messages
-end
-
-get '/clients/:client_id/new_messages' do
-  client = Client.find(params[:client_id])
-  messages = client.messages.where(read_at: nil)
-
-  { messages: messages, status: 200 }.to_json
-end
-
-post '/messages/:message_id/mark_as_read' do
-  message = Message.find(params[:message_id])
-  message.update_attribute(:read_at, Time.now)
-end
-
-get '/error' do
-  erb :error
-end
-
-#settings
-set :public_folder, 'public'
-
-configure :development do
-  set :database, { adapter: "sqlite3", database: "./db/sqlite3" }
-end
-
-configure :production do
- db = URI.parse(ENV['DATABASE_URL'] || 'postgres:///localhost/database')
-
- ActiveRecord::Base.establish_connection(
-   :adapter  => db.scheme == 'postgres' ? 'postgresql' : db.scheme,
-   :host     => db.host,
-   :username => db.user,
-   :password => db.password,
-   :database => db.path[1..-1],
-   :encoding => 'utf8'
- )
-end
